+++ app/src/main/java/com/kau/ttokttok/ui/compose/register/RegisterRoute.kt
@@ -1,10 +1,7 @@
 package com.kau.ttokttok.ui.compose.register
 
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.*
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.kau.ttokttok.ui.component.common.AppDialog
 
@@ -13,8 +10,6 @@ fun RegisterRoute(
     viewModel: RegisterViewModel = hiltViewModel(),
     onSuccess: () -> Unit = { }
 ) {
-    val snackBar = remember { SnackbarHostState() }
-
     val showDialog = remember { mutableStateOf(false) }
     val dialogTitle = remember { mutableStateOf("")}
     val dialogMessage = remember { mutableStateOf("")}
@@ -22,10 +17,6 @@ fun RegisterRoute(
     LaunchedEffect(Unit) {
         viewModel.events.collect { event ->
             when (event) {
-                is RegisterEvent.ShowMessage -> {
-                    snackBar.showSnackbar(event.message)
-                }
-
                 is RegisterEvent.ShowAlert -> {
                     dialogTitle.value = event.title
                     dialogMessage.value = event.message
@@ -43,7 +34,9 @@ fun RegisterRoute(
         AppDialog(
             title = dialogTitle.value,
             message = dialogMessage.value,
-            onDismiss = { showDialog.value = false}
+            onDismiss = {
+                onSuccess()
+            }
         )
     }
