+++ app/src/main/java/com/kau/ttokttok/ui/compose/register/RegisterScreen.kt
@@ -1,25 +1,14 @@
 package com.kau.ttokttok.ui.compose.register
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.AddHome
 import androidx.compose.material.icons.outlined.Email
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -32,10 +21,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.kau.ttokttok.ui.component.auth.AuthInputField
-import com.kau.ttokttok.ui.component.auth.FrostedPanel
-import com.kau.ttokttok.ui.component.auth.NoIconAuthInputField
-import com.kau.ttokttok.ui.component.auth.PasswordField
+import com.kau.ttokttok.ui.component.auth.*
 import com.kau.ttokttok.ui.component.auth.button.RegisterButton
 import com.kau.ttokttok.ui.component.common.background.StarField
 @Preview
